Restructure library modules
Separate helpers into module_utils.helpers module
Add check_query_error helper to check GQL query response
