+++ library/cartridge_probe_server.py
@@ -3,12 +3,8 @@
 import requests
 
 from ansible.module_utils.basic import AnsibleModule
+from ansible.module_utils.helpers import ModuleRes, check_query_error
 
-class ModuleRes:
-    def __init__(self, success, msg=None, changed=False):
-        self.success = success
-        self.msg = msg
-        self.changed = changed
 
 argument_spec = {
     'instance': {'required': True, 'type': 'dict'},
@@ -17,6 +13,7 @@ argument_spec = {
     'control_instance_port': {'required': True, 'type': 'str'},
 }
 
+
 def probe_server(params):
     if 'http_port' not in params['instance']:
        params['instance']['http_port'] = '8080'
@@ -24,8 +21,10 @@ def probe_server(params):
     changed = False
 
     # Get instance URI
-    admin_api_url = 'http://{}:{}/admin/api'.format(params['server_address'], params['instance']['http_port'])
-    print('admin_api_url')
+    admin_api_url = 'http://{}:{}/admin/api'.format(
+        params['server_address'],
+        params['instance']['http_port']
+    )
 
     query = '''
         query {
@@ -38,31 +37,29 @@ def probe_server(params):
     '''
 
     response = requests.post(admin_api_url, json={'query': query})
-    if response.status_code != 200:
-        return ModuleRes(success=False,
-                         msg="Query failed to run by returning code of {}. {}".format(response.status_code, query))
-    if 'errors' in response.json():
-        return ModuleRes(success=False,
-                          msg="Query failed to run with error {}. {}".format(response.json()['errors'][0]['message'], query))
+    err = check_query_error(query, response)
+    if err is not None:
+        return err
 
     instance_uri = response.json()['data']['cluster']['self']['uri']
 
     # Probe instance
     ## NOTE: control instance is used here
-    admin_api_url = 'http://{}:{}/admin/api'.format(params['control_instance_address'], params['control_instance_port'])
+    admin_api_url = 'http://{}:{}/admin/api'.format(
+        params['control_instance_address'],
+        params['control_instance_port']
+    )
     query = '''
         mutation {{
           probe_instance:
             probe_server(uri: "{}")
         }}
     '''.format(instance_uri)
+
     response = requests.post(admin_api_url, json={'query': query})
-    if response.status_code != 200:
-        return ModuleRes(success=False,
-                         msg="Query failed to run by returning code of {}. {}".format(response.status_code, query))
-    if 'errors' in response.json():
-        return ModuleRes(success=False,
-                          msg="Query failed to run with error {}. {}".format(response.json()['errors'][0]['message'], query))
+    err = check_query_error(query, response)
+    if err is not None:
+        return err
 
     probe_success = response.json()['data']['probe_instance']
     return ModuleRes(success=probe_success, changed=changed)
